feat: add kpm update
