+++ kclvm/driver/src/test_data/kpm_update/kcl.mod
@@ -1,0 +1,7 @@
+[package]
+name = "kpm_update"
+edition = "0.0.1"
+version = "0.0.1"
+
+[dependencies]
+k8s = "1.28"
