+++ kclvm/driver/src/tests.rs
@@ -1,4 +1,3 @@
-use super::*;
 use std::path::{Path, PathBuf};
 use std::{env, fs, panic};
 
@@ -8,9 +7,10 @@ use kclvm_parser::LoadProgramOptions;
 use walkdir::WalkDir;
 
 use crate::arguments::parse_key_value_pair;
-use crate::kpm_metadata::{fetch_metadata, fill_pkg_maps_for_k_file, lookup_the_nearest_file_dir};
+use crate::kpm_metadata::fetch_metadata;
 use crate::kpm_update::update_kcl_module;
 use crate::{canonicalize_input_files, expand_input_files, get_pkg_list};
+use crate::{fill_pkg_maps_for_k_file, lookup_the_nearest_file_dir};
 
 #[test]
 fn test_canonicalize_input_files() {
@@ -382,28 +382,16 @@ fn test_get_pkg_list() {
     );
 }
 
-#[cfg(test)]
-// Define a mock structure to simulate the behavior of Command::output
-struct MockCommand {
-    output: Result<MockCommandOutput, std::io::Error>,
-}
-// Define a mock structure to represent the output of Command::output
-struct MockCommandOutput {
-    status: std::process::ExitStatus,
-    stderr: Vec<u8>,
-}
-
 #[test]
-fn test_update_kcl_module_failure() {
-    let manifest_path = PathBuf::from("path/to/manifest");
-    fn mock_command_new_failing(_command: &str) -> MockCommand {
-        MockCommand {
-            output: Err(std::io::Error::new(
-                std::io::ErrorKind::Other,
-                "Command failed",
-            )),
-        }
-    }
-    let result = update_kcl_module(manifest_path);
-    assert!(result.is_err());
+fn test_update_module() {
+    let path = PathBuf::from(".")
+        .join("src")
+        .join("test_data")
+        .join("kpm_update");
+
+    let update_mod = update_kcl_module(path.clone());
+    // Show more information when the test fails.
+    println!("{:?}", update_mod);
+    assert!(!update_mod.is_err());
+    
 }
